Render using carbonplan/maps when pyramid is present
