+++ package.json
@@ -27,6 +27,7 @@
     "@carbonplan/components": "^12.0.0",
     "@carbonplan/icons": "^2.0.0",
     "@carbonplan/layouts": "^3.1.0",
+    "@carbonplan/maps": "^3.1.3",
     "@carbonplan/prism": "^2.0.0",
     "@carbonplan/theme": "^8.0.0",
     "@emotion/react": "^11.7.1",
